% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmap_by_name.R
\name{get_infor}
\alias{get_infor}
\title{Stamen map API}
\usage{
get_infor(name, zoom_l, map_type)
}
\arguments{
\item{name}{city name}

\item{zoom_l}{zoom level}

\item{map_type}{map type}
}
\value{
a ggmap plot
}
\description{
Stamen map API
}
