% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmap_by_name.R
\name{openmap_by_name}
\alias{openmap_by_name}
\title{display a map}
\usage{
openmap_by_name(name, zoom_l, map_type)
}
\arguments{
\item{name}{city name}

\item{zoom_l}{zoom level}

\item{map_type}{map type}
}
\value{
a ggmap plot
}
\description{
display a map
}
